Correct ordering for resources in base class
In the class apache_modjk::mod::jk the concat resources were not ordered
in any way against the required resources in the apache class. This
meant that the apache class may not refresh the httpd service when
mod_jk configuration change.

This commit adds the correct ordering. It also adds default fragments to
the jk_mount and jk_worker concats.